ai-butler/requirements.txt
aiohttp>=3.9.0
edge-tts>=7.0.0
faster-whisper>=1.0.0
numpy>=1.24.0
opencv-python>=4.8.0
PyYAML>=6.0
sounddevice>=0.4.0
